Add error handling for API delete project call
Found that if a project was stuck being deleted, then the next call
from the repear to delete the same project would cause an unhandled
error.  The error would cause the program to terminate without
attempting to process further projects in the list.

This commit adds some error handling so the project is simply
skipped if busy.